More ore stuff

+++ src/main/java/stardust_binding/dimensional_ores/world/DimWorldGenMinable.java
@@ -25,12 +25,12 @@ public class DimWorldGenMinable extends WorldGenerator {
 
     public boolean generate(@Nonnull World worldIn, @Nonnull Random rand, @Nonnull BlockPos position) {
         float f = rand.nextFloat() * (float) Math.PI;
-        double d0 = (double) ((float) (position.getX() + 8) + MathHelper.sin(f) * (float) this.blockCount / 8.0F);
-        double d1 = (double) ((float) (position.getX() + 8) - MathHelper.sin(f) * (float) this.blockCount / 8.0F);
-        double d2 = (double) ((float) (position.getZ() + 8) + MathHelper.cos(f) * (float) this.blockCount / 8.0F);
-        double d3 = (double) ((float) (position.getZ() + 8) - MathHelper.cos(f) * (float) this.blockCount / 8.0F);
-        double d4 = (double) (position.getY() + rand.nextInt(3) - 2);
-        double d5 = (double) (position.getY() + rand.nextInt(3) - 2);
+        double d0 = ((float) (position.getX() + 8) + MathHelper.sin(f) * (float) this.blockCount / 8.0F);
+        double d1 = ((float) (position.getX() + 8) - MathHelper.sin(f) * (float) this.blockCount / 8.0F);
+        double d2 = ((float) (position.getZ() + 8) + MathHelper.cos(f) * (float) this.blockCount / 8.0F);
+        double d3 = ((float) (position.getZ() + 8) - MathHelper.cos(f) * (float) this.blockCount / 8.0F);
+        double d4 = (position.getY() + rand.nextInt(3) - 2);
+        double d5 = (position.getY() + rand.nextInt(3) - 2);
 
         for (int i = 0; i < this.blockCount; ++i) {
             float f1 = (float) i / (float) this.blockCount;
